refactor : App을 컨트롤러로 설정 후 불필요한 파일 제거
