--- src/constant/message.js
@@ -1,9 +1,0 @@
-const GAME_MESSAGE = Object.freeze({
-  CAR_NAMES_INPUT: '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n',
-  ATTEMPT_COUNT_INPUT: '시도할 횟수는 몇 회인가요?\n',
-
-  EXECUTION_RESULT_OUTPUT: '\n실행 결과',
-  WINNER_OUTPUT: '최종 우승자 :',
-});
-
-export default GAME_MESSAGE;
